auto code current version as of 6:22 2/22/16
got robot to run straight and backwards only using encoder values

+++ EndGameCode2016/src/Subsystems/ControlSS.h
@@ -33,7 +33,7 @@ public:
 	//variables to store position and defense values
 	int posNum;
 	int defNum;
-	int afterNum;
+	double afterNum;
 	int windowNum;
 	int loadingPressure;
 	int firingPressure;
